Move files to a separate package
