+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/products/variations/VariationNavigationTarget.kt
@@ -3,10 +3,10 @@ package com.woocommerce.android.ui.products.variations
 import com.woocommerce.android.model.Product.Image
 import com.woocommerce.android.model.SubscriptionDetails
 import com.woocommerce.android.ui.products.ProductInventoryViewModel.InventoryData
-import com.woocommerce.android.ui.products.ProductPricingViewModel.PricingData
 import com.woocommerce.android.ui.products.ProductShippingViewModel.ShippingData
 import com.woocommerce.android.ui.products.SaleDetails
 import com.woocommerce.android.ui.products.models.QuantityRules
+import com.woocommerce.android.ui.products.price.ProductPricingViewModel.PricingData
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event
 
 /**
